requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
scipy==1.11.4
tqdm==4.66.4
matplotlib==3.8.4
scikit-surprise==1.1.4
lightfm==1.17
implicit==0.7.2
gradio==4.36.1
streamlit==1.35.0
requests==2.32.3
pyyaml==6.0.1
